+++ packages/rspress/src/web/ui/ResizablePanels/types.ts
@@ -1,12 +1,9 @@
 import type { ReactElement } from "react";
+import type { ResizablePanelsOptions } from "~shared/types";
 
-export type ResizablePanelsProps = {
+export type ResizablePanelsProps = ResizablePanelsOptions & {
 	editor?: ReactElement;
 	preview?: ReactElement;
-	/**
-	 * Used for auto saving the panel sizes in local storage
-	 */
-	autoSaveId?: string;
 	/**
 	 * Layout width threshold in px.
 	 * When width of the ResizablePanels' wrapper div is smaller,
@@ -15,22 +12,6 @@ export type ResizablePanelsProps = {
 	 * @defaultValue 550
 	 */
 	verticalThreshold?: number;
-	/**
-	 * Default panel sizes.
-	 * Percentage strings (e.g. `"50%"`) or pixel numbers.
-	 */
-	defaultPanelSizes?: {
-		/**
-		 * Default panel size.
-		 * @defaultValue `"50%"`
-		 */
-		editor?: string | number;
-		/**
-		 * Default panel size.
-		 * @defaultValue `"50%"`
-		 */
-		preview?: string | number;
-	};
 	classes?: {
 		wrapper?: string;
 		editorPanel?: string;